fix: rename checkbox component index file to componentRegistry
