+++ components/ui/checkbox/componentRegistry.ts
